requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyascii==0.3.3
arabic-reshaper==3.0.0
blend_modes==2.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
imgaug==0.4.0
Jinja2==3.1.6
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
lmdb==1.5.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.18.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==1.0.0
pytweening==1.2.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
simsimd==6.5.0
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
synthtiger==1.2.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
validators==0.35.0
Werkzeug==3.1.3
